Fixed ordering of vertices in polygons.
Updated vertex ordering on ObjPoly; vertices no longer stored in
ascending order, but in order of setting to allow normal directions
to be retained even if no texture mapping association. This also
entailed changes to ObjModelHoleFiller and ObjModelVertexAdder.
Added debug assertions in ObjModelCurvatureMap when parsing passed
in polygon normal and area objects to ensure that faces in the source
model are referenced.
Fixed bug in ObjModelComponentFinder which was ignoring components
having no boundary edges if there was only a single boundary (on
one component) for the whole model.