ObjModel and ObjModelRegionSelector changes.
Added calculation of polygon normals in ObjModel and
calculation of basis vectors for a polygon based on
maintained ordering of vertices in polygons.

Changed interface of ObjModelRegionSelector to use
offsets to a referenced vertex index stored in a
local coordinate frame so that when the referenced
ObjModel is transformed, centre of region remains
fixed with respect to the model.

Changed interface to ObjModelComponentFinder
to return index to component (componentFromVertex).